+++ src/Propulsion.CosmosStore/CosmosStoreSink.fs
@@ -12,6 +12,17 @@ open System.Collections.Generic
 [<AutoOpen>]
 module private Impl =
     let inline mb x = float x / 1024. / 1024.
+#if COSMOSV3 || COSMOSV2
+    // The event bodies the store deals with are byte arrays
+    let toNativeEvent = id
+#else
+    // v4 and later use JsonElement, but Propulsion is sticking with byte arrays until 3.x (at which point it'll probably shift to ReadOnlyMemory<byte> rather than assuming and/or offering optimization for JSON bodies)
+    open System.Text.Json
+    let toNativeEventBody (x : byte[]) : JsonElement =
+        if x = null then JsonElement()
+        else JsonSerializer.Deserialize(System.ReadOnlySpan.op_Implicit x)
+    let toNativeEvent = FsCodec.Core.TimelineEvent.Map toNativeEventBody
+#endif
 
 module Internal =
 
@@ -40,7 +51,7 @@ module Internal =
 
         let write (log : ILogger) (ctx : EventsContext) stream span = async {
             log.Debug("Writing {s}@{i}x{n}", stream, span.index, span.events.Length)
-            let! res = ctx.Sync(stream, { index = span.index; etag = None }, span.events |> Array.map (fun x -> x :> _))
+            let! res = ctx.Sync(stream, { index = span.index; etag = None }, span.events |> Array.map (fun x -> toNativeEvent x :> _))
             let res' =
                 match res with
                 | AppendResult.Ok pos -> Ok pos.index