+++ src/Propulsion.Feed/FeedSource.fs
@@ -9,20 +9,18 @@ open System
 /// Drives reading and checkpointing for a set of feeds (tranches) of a custom source feed
 type FeedSourceBase internal
     (   log : Serilog.ILogger, statsInterval : TimeSpan, sourceId,
-        checkpoints : IFeedCheckpointStore, defaultCheckpointEventInterval : TimeSpan,
+        checkpoints : IFeedCheckpointStore, establishOrigin : (TrancheId -> Async<Position>) option,
         sink : ProjectorPipeline<Ingestion.Ingester<seq<StreamEvent<byte[]>>, Submission.SubmissionBatch<int,StreamEvent<byte[]>>>>) =
 
-    let log =
-        log.ForContext("instanceId", let g = Guid.NewGuid() in g.ToString "N")
-           .ForContext("source", sourceId)
+    let log = log.ForContext("source", sourceId)
 
     let pumpPartition crawl partitionId trancheId = async {
         let log = log.ForContext("tranche", trancheId)
         let ingester : Ingestion.Ingester<_, _> = sink.StartIngester(log, partitionId)
         let reader = FeedReader(log, sourceId, trancheId, statsInterval, crawl trancheId, ingester.Submit, checkpoints.Commit)
-        try let! freq, pos = checkpoints.Start(sourceId, trancheId, defaultCheckpointEventInterval)
-            log.Information("Reading {sourceId}/{trancheId} @ {pos} checkpointing every {checkpointFreq:n1}m", sourceId, trancheId, pos, freq.TotalMinutes)
-            do! reader.Pump(pos)
+        try let! freq, pos = checkpoints.Start(sourceId, trancheId, ?establishOrigin = (match establishOrigin with None -> None | Some f -> Some (f trancheId)))
+            log.Information("Reading {source:l}/{tranche:l} From {pos} Checkpoint Event interval {checkpointFreq:n1}m", sourceId, trancheId, pos, freq.TotalMinutes)
+            return! reader.Pump(pos)
         with e ->
             log.Warning(e, "Exception encountered while running reader, exiting loop")
             return! Async.Raise e
@@ -31,16 +29,57 @@ type FeedSourceBase internal
     /// Propagates exceptions raised by <c>readTranches</c> or <c>crawl</c>,
     member internal _.Pump
         (   readTranches : unit -> Async<TrancheId[]>,
-            /// Responsible for managing retries and back offs; yielding an exception will result in abend of the read loop
-            crawl : TrancheId -> bool * Position -> AsyncSeq<Internal.Batch<byte[]>>) = async {
+            // Responsible for managing retries and back offs; yielding an exception will result in abend of the read loop
+            crawl : TrancheId -> bool * Position -> AsyncSeq<TimeSpan * Batch<byte[]>>) = async {
         try let! tranches = readTranches ()
             log.Information("Starting {tranches} tranche readers...", tranches.Length)
-            let crawl trancheId (wasLast, pos) = asyncSeq {
-                yield! crawl trancheId (wasLast, pos) }
             return! Async.Parallel(tranches |> Seq.mapi (pumpPartition crawl)) |> Async.Ignore<unit[]>
         with e ->
             log.Warning(e, "Exception encountered while running source, exiting loop")
             return! Async.Raise e }
+
+/// Drives reading and checkpointing from a source that contains data from multiple streams. While a TrancheId is always required,
+/// it may have a default value of `"0"` if the underlying source representation does not involve autonomous shards/physical partitions etc
+type TailingFeedSource
+    (   log : Serilog.ILogger, statsInterval : TimeSpan,
+        sourceId, tailSleepInterval : TimeSpan,
+        crawl : TrancheId * Position -> AsyncSeq<TimeSpan * Batch<_>>,
+        checkpoints : IFeedCheckpointStore, establishOrigin : (TrancheId -> Async<Position>) option,
+        sink : ProjectorPipeline<Ingestion.Ingester<seq<StreamEvent<byte[]>>, Submission.SubmissionBatch<int,StreamEvent<byte[]>>>>) =
+    inherit FeedSourceBase(log, statsInterval, sourceId, checkpoints, establishOrigin, sink)
+
+    let crawl trancheId (wasLast, startPos) = asyncSeq {
+        if wasLast then
+            do! Async.Sleep tailSleepInterval
+        try let batches = crawl (trancheId, startPos)
+            for batch in batches do
+                yield batch
+        with e ->
+            log.Warning(e, "Read failure")
+            do! Async.Sleep tailSleepInterval }
+
+    member _.Pump(readTranches) =
+        base.Pump(readTranches, crawl)
+
+/// Drives reading and checkpointing from a source that aggregates data from multiple streams as a singular source
+/// without shards/physical partitions (tranches), such as the SqlStreamStore, and EventStoreDB $all feeds
+/// Per the API design of such stores, readBatch only ever yields a single page
+type AllFeedSource
+    (   log : Serilog.ILogger, statsInterval : TimeSpan,
+        sourceId, tailSleepInterval : TimeSpan,
+        readBatch : Position -> Async<Batch<_>>,
+        checkpoints : IFeedCheckpointStore,
+        sink : ProjectorPipeline<Ingestion.Ingester<seq<StreamEvent<byte[]>>, Submission.SubmissionBatch<int,StreamEvent<byte[]>>>>) =
+    inherit TailingFeedSource(log, statsInterval, sourceId, tailSleepInterval,
+                              (fun (_trancheId, pos) -> asyncSeq {
+                                                            let sw = System.Diagnostics.Stopwatch.StartNew()
+                                                            let! b = readBatch pos
+                                                            yield sw.Elapsed, b } ),
+                              checkpoints, None, sink)
+
+    member _.Pump() =
+        let readTranches () = async { return [| TrancheId.parse "0" |] }
+        base.Pump(readTranches)
 
 namespace Propulsion.Feed
 
@@ -58,21 +97,21 @@ type Page<'e> = { items : FsCodec.ITimelineEvent<'e>[]; checkpoint : Position; i
 type FeedSource
     (   log : Serilog.ILogger, statsInterval : TimeSpan,
         sourceId, tailSleepInterval : TimeSpan,
-        checkpoints : IFeedCheckpointStore, defaultCheckpointEventInterval : TimeSpan,
-        /// Responsible for managing retries and back offs; yielding an exception will result in abend of the read loop
+        checkpoints : IFeedCheckpointStore,
+        // Responsible for managing retries and back offs; yielding an exception will result in abend of the read loop
         readPage : TrancheId * Position -> Async<Page<byte[]>>,
         sink : ProjectorPipeline<Ingestion.Ingester<seq<StreamEvent<byte[]>>, Submission.SubmissionBatch<int,StreamEvent<byte[]>>>>) =
-    inherit Internal.FeedSourceBase(log, statsInterval, sourceId, checkpoints, defaultCheckpointEventInterval, sink)
+    inherit Internal.FeedSourceBase(log, statsInterval, sourceId, checkpoints, None, sink)
 
     let crawl trancheId =
         let streamName = FsCodec.StreamName.compose "Messages" [SourceId.toString sourceId; TrancheId.toString trancheId]
         fun (wasLast, pos) -> asyncSeq {
             if wasLast then
                 do! Async.Sleep tailSleepInterval
+            let sw = System.Diagnostics.Stopwatch.StartNew()
             let! page = readPage (trancheId, pos)
             let items' = page.items |> Array.map (fun x -> { stream = streamName; event = x })
-            let slice : Internal.Batch<_> = { items = items'; checkpoint = page.checkpoint; isTail = page.isTail }
-            yield slice
+            yield sw.Elapsed, ({ items = items'; checkpoint = page.checkpoint; isTail = page.isTail } : Internal.Batch<_>)
         }
 
     /// Drives the continual loop of reading and checkpointing each tranche until a fault occurs. <br/>