+++ src/Propulsion.SqlStreamStore/ReaderCheckpoint.fs
@@ -44,18 +44,21 @@ let commitPosition (conn : IDbConnection) (stream : string) (consumerGroup : str
      |> Async.AwaitTaskCorrect
      |> Async.Ignore<int>
 
-let getPosition (conn : IDbConnection) (stream : string) (consumerGroup : string) = async {
+let tryGetPosition (conn : IDbConnection) (stream : string) (consumerGroup : string) = async {
     let! res =
         conn.QueryAsync<CheckpointEntry>(
             """SELECT * FROM Checkpoints WHERE Stream = @Stream AND ConsumerGroup = @ConsumerGroup""",
             { Stream = stream; ConsumerGroup = consumerGroup; Position = Nullable() })
         |> Async.AwaitTaskCorrect
 
-    match Seq.tryHead res with
-    | Some res -> return res.Position
-    | None -> return Nullable() }
+    return Seq.tryHead res |> Option.bind (fun r -> Option.ofNullable r.Position) }
 
-type Service(connString : string) =
+type Service(connString : string, consumerGroupName, defaultCheckpointFrequency) =
+
+    let streamName source tranche =
+        match SourceId.toString source, TrancheId.toString tranche with
+        | s, null -> s
+        | s, tid -> String.Join("_", s, tid)
 
     member _.CreateSchemaIfNotExists() = async {
         use conn = createConnection connString
@@ -63,11 +66,16 @@ type Service(connString : string) =
 
     interface IFeedCheckpointStore with
 
-        member _.Start(source, tranche, defaultCheckpointFrequency) = async {
+        member _.Start(source, tranche, ?establishOrigin) = async {
             use conn = createConnection connString
-            let! pos = getPosition conn (SourceId.toString source) (TrancheId.toString tranche)
-            return defaultCheckpointFrequency, if pos.HasValue then Position.parse pos.Value else Position.initial }
+            let! maybePos = tryGetPosition conn (streamName source tranche) consumerGroupName
+            let! pos =
+                match maybePos, establishOrigin with
+                | Some pos, _ -> async { return Position.parse pos }
+                | None, Some f -> f
+                | None, None -> async { return Position.initial }
+            return defaultCheckpointFrequency, pos }
 
         member _.Commit(source, tranche, pos) = async {
             use conn = createConnection connString
-            return! commitPosition conn (SourceId.toString source) (TrancheId.toString tranche) (Position.toInt64 pos) }
+            return! commitPosition conn (streamName source tranche) consumerGroupName (Position.toInt64 pos) }